pyproject.toml
[project]
name = "myapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "djangorestframework-jwt>=1.11.0",
    "pillow>=11.1.0",
    "django-cors-headers>=4.6.0",
    "python-decouple>=3.8",
    "dj-database-url>=2.3.0",
]
